feat: add node only server support and example
