+++ examples/node-server-only/config/config.js
@@ -1,0 +1,3 @@
+exports.default = {
+    NODE_ONLY: true,
+}
