+++ examples/node-server-only/server/start.ts
@@ -1,0 +1,17 @@
+import { createNodeServer } from 'lib/runtime/node-server'
+import { consoleLogger } from 'typescript-log'
+
+const log = consoleLogger()
+
+createNodeServer({
+    log,
+    middlewareHook: app => {
+        app.get('/', (_req, res) => {
+            res.send('The magic is real')
+        })
+
+        app.get('/another', (_req, res) => {
+            res.send('another endpoint')
+        })
+    },
+})
