+++ lib/runtime/server/node-server.ts
@@ -1,0 +1,5 @@
+import express from 'express'
+import { createCustomServer, CreateServerOptions } from './custom-server'
+
+export const createNodeServer = (options: CreateServerOptions): express.Express =>
+    createCustomServer({ ...options, nodeOnlyServer: true })
