+++ lib/stats/server.ts
@@ -4,7 +4,7 @@ import { BuildConfig } from 'lib/runtime/server'
 import os from 'os'
 import path from 'path'
 import { Logger } from 'typescript-log'
-import { getPort } from '../runtime/server/server'
+import { getPort } from '../runtime/server/custom-server'
 import { findFreePort, waitForConnection } from '../runtime/util/network'
 import { forkPromise } from '../util/process'
 import { getTimeMs, timeout } from '../util/time'
