+++ package.json
@@ -94,6 +94,10 @@
         "watch:emotion-css-in-js-example": "yarn pwt watch server -p examples/emotion-css-in-js",
         "stats:emotion-css-in-js-example": "yarn pwt stats -p examples/emotion-css-in-js",
         "explore-bundle:emotion-css-in-js-example": "yarn pwt explore-bundle -p examples/emotion-css-in-js",
+        "~~~ node-server-only ~~~": "",
+        "build:node-server-only": "yarn pwt build -p examples/node-server-only",
+        "watch:node-server-only": "yarn pwt watch server -p examples/node-server-only",
+        "explore-bundle:node-server-only": "yarn pwt explore-bundle -p examples/node-server-only",
         "~~~ ssr-with-sass ~~~": "",
         "prebuild:ssr-with-sass-example": "yarn clean:lighthouse && yarn build",
         "prewatch:ssr-with-sass-example": "yarn clean:lighthouse && yarn build",
